Unwrap responses and return native error types
