+++ go.work
@@ -1,2 +1,4 @@
 go 1.20
-go work use ./go/ ./languages/go/
+
+use ./languages/go
+use ./languages/go/example
